Refactor Stage-Gate Validator and Update Project Structure
- Updated validation criteria in stage_gate_validator.py to align with project specifications.
- Simplified validation execution flow and improved report generation.
- Removed outdated test files for EfficientNetV2 and model selection.
- Enhanced dataset handling in training scripts with MultiDatasetWrapper.
- Updated training configuration to use EfficientNetV2-B3 and adjusted class count handling.
- Added new README files for datasets and models, outlining structure and usage.
- Introduced environment manager script for automated PyTorch installation based on GPU compatibility.
- Comprehensive updates to testing framework, including new dataset tests and experiment management utilities.

+++ src/stage_00/dataset.py
@@ -142,8 +142,8 @@ class CelebDFDataset(Dataset):
             transformed = self.transform(image=image)
             image_tensor = transformed['image']
             
-            # Convert label to tensor
-            label_tensor = torch.tensor(label, dtype=torch.long)
+            # Convert label to tensor - use float for BCE compatibility
+            label_tensor = torch.tensor(label, dtype=torch.float)
             
             return image_tensor, label_tensor
             
@@ -151,7 +151,7 @@ class CelebDFDataset(Dataset):
             print(f"Error loading image {image_path}: {str(e)}")
             # Return a blank image and label
             blank_image = torch.zeros(3, self.image_size, self.image_size)
-            return blank_image, torch.tensor(0, dtype=torch.long)
+            return blank_image, torch.tensor(0, dtype=torch.float)
     
     def get_class_weights(self) -> torch.Tensor:
         """
